pyproject.toml
[tool.poetry]
name = "kdv"
version = "0.1.0"
description = ""
authors = ["Rasen Nikov <rasen800@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-bootstrap5 = "^23.4"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
